% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_save_dbi_output.R
\name{clean_save_dbi_output}
\alias{clean_save_dbi_output}
\title{Clean DBI output and save as RDA file
Last updated: 2021-03-08}
\usage{
clean_save_dbi_output(r_covariates, r_drugs, r_biomarkers)
}
\arguments{
\item{r_covariates}{A tibble}

\item{r_drugs}{A tibble,}

\item{r_biomarkers}{A tibble}
}
\description{
Clean DBI output and save as RDA file
Last updated: 2021-03-08
}
