% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_netezza_queries_v2.R
\name{create_netezza_queries_v2}
\alias{create_netezza_queries_v2}
\title{Function to create covariates table  over DBI with javaODBC for Netezza,with option to set ancestor concept_id table name
Last updated: 2021-03-09}
\usage{
create_netezza_queries_v2(
  drug_concept_id,
  biomarker_concept_id,
  table_name,
  table_type,
  ancestor_table,
  conn,
  ...
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{table_name}{A string}

\item{table_type}{A string, either c("covariates", "drugs", "biomarkers")}

\item{ancestor_table}{A string, eg, "antihypertensive_drugs_ancestor_tbl_20210308151236"}

\item{conn}{A variable}
}
\description{
Function to create covariates table  over DBI with javaODBC for Netezza,with option to set ancestor concept_id table name
Last updated: 2021-03-09
}
