% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_word_table.R
\name{create_word_table}
\alias{create_word_table}
\title{Use flextable package to create table for word docx
Last updated: 2021-03-02}
\usage{
create_word_table(file_name, df, file_type = "excel", pvalue_col = T)
}
\arguments{
\item{file_name}{A string, e.g., "ddiwas_supptable3_ppv_table_20200914085357.xlsx"}

\item{df}{A tibble. Specify if using tibble instead of file name}

\item{file_type}{A string, e.g., "excel" or "csv"}

\item{pvalue_col}{A boolean. Default = T. Set = F if there is not a pvalue column}
}
\description{
Use flextable package to create table for word docx
Last updated: 2021-03-02
}
