% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cohort_table.R
\name{generate_cohort_table}
\alias{generate_cohort_table}
\title{For drug repurposing EHR validation studies, generates demographics table
Last updated: 2020-12-25}
\usage{
generate_cohort_table(
  t_phenotypes_variables,
  cohort_drug,
  drug_cohort_metadata,
  col_vector
)
}
\arguments{
\item{cohort_drug}{A string}

\item{drug_cohort_metadata}{A tibble}

\item{col_vector}{A vector}

\item{t_phenotype_variables}{A tibble}
}
\description{
For drug repurposing EHR validation studies, generates demographics table
Last updated: 2020-12-25
}
