% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cohort_selection_numbers.R
\name{get_cohort_selection_numbers}
\alias{get_cohort_selection_numbers}
\title{Get cohort selection numbers for drug repurposing}
\usage{
get_cohort_selection_numbers(
  drug,
  phenotype,
  biomarker,
  indication_drugs,
  r_covariates,
  r_drugs,
  r_biomarkers
)
}
\arguments{
\item{drug}{A string, e.g., "simvastatin"}

\item{phenotype}{A string, e.g., "hyperlipidemia"}

\item{biomarker}{A string, e.g., "LDL cholesterol"}

\item{indication_drugs}{A tibble with columns c('rxcui_in', 'drug_desc', 'indication')}

\item{r_covariates}{A tibble, with columns c('person_id','dob','gender_concept_id','race_concept_id','drug_concept_id','first_drug_exposure','last_drug_exposure')}

\item{r_drugs}{A tibble, with columns c('person_id','start_date','end_date','drug_exposure_start_date','drug_concept_id')}

\item{r_biomarkers}{A tibble, with columns c('person_id','start_date','measurement_date','end_date','measurement_concept_id','value_as_number')}
}
\description{
Get cohort selection numbers for drug repurposing
}
\details{
Last updated: 2021-10-15
}
