% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netezza_query_icds.R
\name{netezza_query_icds}
\alias{netezza_query_icds}
\title{Function to get ICD data for calculating elixhauser scores
Updated: 2021-10-01}
\usage{
netezza_query_icds(
  drug_concept_id,
  biomarker_concept_id,
  table_name,
  table_type,
  ancestor_table,
  conn,
  ...
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{table_name}{A string}

\item{table_type}{A string, either c("covariates", "drugs", "biomarkers", "icds")}

\item{ancestor_table}{A string, eg, "antihypertensive_drugs_ancestor_tbl_20210308151236"}

\item{conn}{A variable}
}
\description{
Function to get ICD data for calculating elixhauser scores
Updated: 2021-10-01
}
