% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repurposing_query_biomarkers.R
\name{repurposing_query_biomarkers}
\alias{repurposing_query_biomarkers}
\title{Function to create SQL query for drugs data
Last updated: 2021-02-15}
\usage{
repurposing_query_biomarkers(
  drug_concept_id,
  biomarker_concept_id,
  upper_bound_biomarker,
  db = "netezza",
  ...
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{upper_bound_biomarker}{A string}

\item{db}{A string, either "netezza" or "bigquery"}
}
\description{
Function to create SQL query for drugs data
Last updated: 2021-02-15
}
