% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repurposing_query_covariates.R
\name{repurposing_query_covariates}
\alias{repurposing_query_covariates}
\title{Function to create SQL query for covariates data
Last updated: 2021-02-15}
\usage{
repurposing_query_covariates(
  drug_concept_id,
  biomarker_concept_id,
  db = "netezza",
  ...
)
}
\arguments{
\item{drug_concept_id}{A string}

\item{biomarker_concept_id}{A string}

\item{db}{A string, either "netezza" or "bigquery"}
}
\description{
Function to create SQL query for covariates data
Last updated: 2021-02-15
}
