% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_results_table.R
\name{show_results_table}
\alias{show_results_table}
\title{Show results table}
\usage{
show_results_table(data_table, length = 100)
}
\arguments{
\item{data_table}{A tibble}

\item{length}{A numeric}
}
\description{
Show results table
}
\details{
Last updated: 2021-03-29
}
